Release VPNExample 13.0.0
### Breaking

- use Swift Package Manager dependencies
- require iOS 18 minimum version

### Fix

- update dependencies

+++ vpn/VPNExample/Models/EntitlementConsumptionModel.swift
@@ -12,8 +12,8 @@ struct EntitlementConsumptionModel: Equatable {
     // MARK: Properties
 
     var name: String
-    var value: Int
-    var consumed: Int
-    var available: Int
+    var value: Int64
+    var consumed: Int64
+    var available: Int64
 
 }
